一部のViewModelをViewの拡張に変更
