+++ TodoList_SwiftUI/Screens/TagScreens/TagListView/TagListView+ViewModel.swift
@@ -1,0 +1,22 @@
+//
+//  TagListViewModel.swift
+//  TodoList_SwiftUI
+//
+//  Created by 土橋正晴 on 2023/07/10.
+//
+
+import Foundation
+
+extension TagListView {
+    
+    class TagListViewModel: ObservableObject {
+        
+        @Published var model: [Tag] = []
+            
+        @MainActor
+        func fetchAllTag() async {
+            model = Tag.findAll()
+        }
+        
+    }
+}
